requirements.txt

fastapi
uvicorn
sqlalchemy
pandas
openpyxl
python-multipart
loguru
python-dotenv
openai
passlib[bcrypt]
python-jose[cryptography]

google-generativeai
flask
flask_cors
escavador
python-docx
